Added refactoring, step 1, stable
