+++ ss_backend/internal/stocksdata/serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from internal.stocksdata.models import StockData
+
+class StockDataSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = StockData
+        fields = '__all__'
+
