+++ ss_backend/internal/symbol/serializers.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from internal.symbol.models import Symbol
+
+class SymbolSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Symbol
+        fields = '__all__'
