+++ ss_backend/internal/watchlists/serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from internal.watchlists.models import WatchList
+
+
+class WatchListSerializer(serializers.ModelSerializer):
+    class Meta:
+        model= WatchList
+        fields ='__all__'
