+++ ss_backend/internal/watchlists/urls.py
@@ -1,0 +1,9 @@
+from django.db import router
+from django.urls import path, include
+from internal.watchlists.views import WatchListView
+
+
+urlpatterns = [
+    path('watchlists/',WatchListView.as_view(),name='watchlists'),
+]
+
